cleanup

+++ Library/MarqueeLabel/MarqueeLabel.m
@@ -1225,7 +1225,7 @@ CGPoint MLOffsetCGPoint(CGPoint point, CGFloat offset);
     }
     
     // Do not allow negative values
-    _leadingBuffer = fabsf(leadingBuffer);
+    _leadingBuffer = fabs(leadingBuffer);
     [self updateSublabelAndLocations];
 }
 
@@ -1235,7 +1235,7 @@ CGPoint MLOffsetCGPoint(CGPoint point, CGFloat offset);
     }
     
     // Do not allow negative values
-    _trailingBuffer = fabsf(trailingBuffer);
+    _trailingBuffer = fabs(trailingBuffer);
     [self updateSublabelAndLocations];
 }
